# ⚡ Optimized `Encoder.EncodeDigits()`