Tested with 1,000,000 `long`s. Hot path for `Binary`, `Text` encoding.
Neither requires string conversion nor additional memory allocation. 😗

As for `Encoder.EncodeInteger()`, conversion from `BigInteger` to `long` seems heaver than `ToString()`, so it is left as it is. 😕

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.3086)
11th Gen Intel Core i7-1185G7 3.00GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.203
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2


|   Method |     Mean |    Error |   StdDev |       Gen0 | Allocated |
|--------- |---------:|---------:|---------:|-----------:|----------:|
| GetBytes | 64.25 ms | 1.320 ms | 3.786 ms | 17800.0000 | 110.53 MB |
|  ForLoop | 42.02 ms | 0.840 ms | 2.396 ms |          - |   3.81 MB |